# Backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Authentication & App API",
    "version": "1.0.0",
    "description": "Production-ready API documentation for Frontend integration"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token without the 'Bearer ' prefix"
      }
    }
  },
  "paths": {
    "/register": {
      "post": {
        "summary": "Register a new user",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userName",
                  "email",
                  "password"
                ],
                "properties": {
                  "userName": {
                    "type": "string",
                    "example": "Rahul123"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "rahul@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "SecurePass123!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Invalid input data / Validation failed"
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Login existing user",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "rahul@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "SecurePass123!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful. Returns access and refresh tokens."
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/refresh": {
      "post": {
        "summary": "Generate new access and refresh tokens",
        "description": "Expects a valid refresh token in the cookies or body.",
        "tags": [
          "Authentication"
        ],
        "responses": {
          "200": {
            "description": "New tokens generated successfully"
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        }
      }
    },
    "/logout": {
      "post": {
        "summary": "Logout current user",
        "description": "Clears cookies or access tokens from session.",
        "tags": [
          "Authentication"
        ],
        "responses": {
          "200": {
            "description": "Logged out successfully"
          }
        }
      }
    },
    "/email-verification": {
      "post": {
        "summary": "verify user email",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": "-email",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "userexample@email.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User email verify successfully"
          },
          "400": {
            "description": "Invalid user email"
          }
        }
      }
    },
    "/verify-email": {
      "get": {
        "summary": "Verify user email using token",
        "description": "The link clicked from user email lands here.",
        "tags": [
          "Authentication"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The verification token sent via email"
          }
        ],
        "responses": {
          "200": {
            "description": "Email verified successfully"
          },
          "400": {
            "description": "Invalid or expired token"
          }
        }
      }
    },
    "/user/{userId}/avatar": {
      "post": {
        "summary": "Upload user profile avatar image",
        "tags": [
          "User Avatar"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the user"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image file to upload as avatar"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Avatar uploaded successfully"
          },
          "400": {
            "description": "Invalid file format or missing file"
          }
        }
      },
      "put": {
        "summary": "Update or replace existing user avatar",
        "tags": [
          "User Avatar"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the user"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary",
                    "description": "New image file to replace the old avatar"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Avatar updated successfully"
          },
          "400": {
            "description": "Update failed due to invalid file type"
          }
        }
      },
      "delete": {
        "summary": "Delete user avatar image",
        "tags": [
          "User Avatar"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "Avatar removed successfully"
          },
          "404": {
            "description": "Avatar not found for this user"
          }
        }
      }
    },
    "/user/all": {
      "get": {
        "summary": "Fetch all registered user accounts",
        "tags": [
          "User Account"
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched list of all accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/create-profile": {
      "post": {
        "summary": "Save or update full user profile details",
        "tags": [
          "User Account"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "firstName",
                  "lastName",
                  "DOB"
                ],
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "Rahul"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Sharma"
                  },
                  "DOB": {
                    "type": "Date",
                    "example": "2004-02-04"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile Details successfully saved or updated"
          },
          "400": {
            "description": "Profile data failed schema validations"
          },
          "401": {
            "description": "Unauthorized. Missing or invalid access token."
          }
        }
      }
    },
    "/user/{userId}": {
      "delete": {
        "summary": "Permanently delete a user account",
        "tags": [
          "User Account"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the account to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Account successfully deleted from the database"
          },
          "401": {
            "description": "Unauthorized access token"
          },
          "403": {
            "description": "Forbidden. You do not have permissions to delete this account."
          }
        }
      }
    }
  },
  "tags": []
}